refactor(stt): split stt_service.py into focused sub-modules (Related to #1938)
Extract the 523-line monolith into four single-responsibility modules:

- stt/pinyin.py          — PINYIN_GROUPS map + get_pinyin() + is_homophone()
- stt/normalization.py   — normalize_numbers() + normalize_for_comparison()
- stt/algorithm.py       — correct_homophones() + compute_match_rate()
- stt/evaluate.py        — evaluate_reading(), Tier, FEEDBACK_KEYS

stt_service.py becomes a thin facade re-exporting all public symbols for
backward compatibility (reading_evaluation_service.py callers unchanged).

TDD: 21 characterization tests in test_stt_refactored.py cover per-module
imports, behaviour parity, and facade backward-compat.
All 40 tests pass (19 original + 21 new).
